first hook from App comp
