+++ resources/js/components/controls/hooks/index.js
@@ -1,0 +1,1 @@
+export * from './misc';
